refactor(cryptox): consolidate and reorganize test files for standard naming
- Merged coverage-specific test files into main test files:
  - aes_100_coverage_test.go → aes_test.go
  - des_100_coverage_test.go → des_test.go
  - rsa_100_coverage_test.go → rsa_test.go
  - ecc_100_coverage_test.go split between ecdh_test.go and ecdsa_test.go
- Created hash_test.go by merging all hash-related tests:
  - hash_basic_test.go, hash_crc_test.go, hash_fnv_test.go, hash_hmac_test.go
- Integrated missing_coverage_test.go content into appropriate files
- Removed 9 separate test files, now 7 files follow {source}_test.go pattern
- Combined unit tests and benchmark tests in single files
- Maintains comprehensive test coverage with cleaner structure
