--- cryptox/ecc_100_coverage_test.go
@@ -1,2324 +1,0 @@
-package cryptox
-
-import (
-	"crypto/ecdsa"
-	"crypto/elliptic"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/sha256"
-	"crypto/x509"
-	"encoding/pem"
-	"errors"
-	"math/big"
-	"reflect"
-	"strings"
-	"testing"
-	"unsafe"
-)
-
-// Mock failures for ECC dependency injection
-type FailingECCReader struct{}
-
-func (fr FailingECCReader) Read(p []byte) (n int, err error) {
-	return 0, errors.New("simulated random reader failure")
-}
-
-// TestECC_100PercentCoverage tests all error paths in ECC functions
-func TestECC_100PercentCoverage(t *testing.T) {
-	// Save original functions
-	originalECDSARandReader := ecdsaRandReader
-	originalECDHRandReader := ecdhRandReader
-
-	// Restore original functions after test
-	defer func() {
-		ecdsaRandReader = originalECDSARandReader
-		ecdhRandReader = originalECDHRandReader
-	}()
-
-	// Test 1: Trigger rand.Reader failure in ECDSA key generation
-	ecdsaRandReader = FailingECCReader{}
-	ecdhRandReader = originalECDHRandReader
-
-	_, err := GenerateECDSAKey(elliptic.P256())
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDSAKey")
-	}
-
-	_, err = GenerateECDSAP256Key()
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDSAP256Key")
-	}
-
-	_, err = GenerateECDSAP384Key()
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDSAP384Key")
-	}
-
-	_, err = GenerateECDSAP521Key()
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDSAP521Key")
-	}
-
-	// Test 2: Trigger rand.Reader failure in ECDH key generation
-	ecdsaRandReader = originalECDSARandReader
-	ecdhRandReader = FailingECCReader{}
-
-	_, err = GenerateECDHKey(elliptic.P256())
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDHKey")
-	}
-
-	_, err = GenerateECDHP256Key()
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDHP256Key")
-	}
-
-	_, err = GenerateECDHP384Key()
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDHP384Key")
-	}
-
-	_, err = GenerateECDHP521Key()
-	if err == nil {
-		t.Error("Expected rand.Reader error in GenerateECDHP521Key")
-	}
-
-	// Test 3: Trigger rand.Reader failure in ECDSA signing
-	ecdsaRandReader = FailingECCReader{}
-	ecdhRandReader = originalECDHRandReader
-
-	// Generate a valid key first with original reader
-	ecdsaRandReader = originalECDSARandReader
-	keyPair, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate test key: %v", err)
-	}
-
-	// Now set failing reader and test signing
-	ecdsaRandReader = FailingECCReader{}
-
-	_, _, err = ECDSASign(keyPair.PrivateKey, []byte("test"), sha256.New)
-	if err == nil {
-		t.Error("Expected rand.Reader error in ECDSASign")
-	}
-
-	_, _, err = ECDSASignSHA256(keyPair.PrivateKey, []byte("test"))
-	if err == nil {
-		t.Error("Expected rand.Reader error in ECDSASignSHA256")
-	}
-
-	_, _, err = ECDSASignSHA512(keyPair.PrivateKey, []byte("test"))
-	if err == nil {
-		t.Error("Expected rand.Reader error in ECDSASignSHA512")
-	}
-
-	// Restore readers
-	ecdsaRandReader = originalECDSARandReader
-	ecdhRandReader = originalECDHRandReader
-}
-
-// TestECDSAPEMErrorPaths tests PEM parsing error paths
-func TestECDSAPEMErrorPaths(t *testing.T) {
-	// Test invalid PEM types
-	invalidPrivateKeyPEM := `-----BEGIN RSA PRIVATE KEY-----
-MIIEpAIBAAKCAQEA1234567890
------END RSA PRIVATE KEY-----`
-
-	_, err := ECDSAPrivateKeyFromPEM([]byte(invalidPrivateKeyPEM))
-	if err == nil {
-		t.Error("Expected error for wrong PEM type")
-	}
-
-	invalidPublicKeyPEM := `-----BEGIN RSA PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA1234567890
------END RSA PUBLIC KEY-----`
-
-	_, err = ECDSAPublicKeyFromPEM([]byte(invalidPublicKeyPEM))
-	if err == nil {
-		t.Error("Expected error for wrong PEM type")
-	}
-
-	// Test valid PEM type but invalid key data
-	invalidPrivateKeyData := `-----BEGIN EC PRIVATE KEY-----
-aW52YWxpZCBkYXRh
------END EC PRIVATE KEY-----`
-
-	_, err = ECDSAPrivateKeyFromPEM([]byte(invalidPrivateKeyData))
-	if err == nil {
-		t.Error("Expected error for invalid private key data")
-	}
-
-	invalidPublicKeyData := `-----BEGIN PUBLIC KEY-----
-aW52YWxpZCBkYXRh
------END PUBLIC KEY-----`
-
-	_, err = ECDSAPublicKeyFromPEM([]byte(invalidPublicKeyData))
-	if err == nil {
-		t.Error("Expected error for invalid public key data")
-	}
-
-	// Test public key with wrong key type - create a malformed PEM with correct header but invalid data
-	malformedPublicKey := `-----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA7u25K9xP8lmE9v3U8OVt
-w0N6HVc5tNOIgbzLFfvO8mWWnhxF11gK2JvKm9lGqLpbOy1rGtCNHsFGj5kD7MQX
------END PUBLIC KEY-----`
-
-	_, err = ECDSAPublicKeyFromPEM([]byte(malformedPublicKey))
-	if err == nil {
-		t.Error("Expected error for non-ECDSA public key")
-	}
-}
-
-// TestECDSASignatureBytesErrorPaths tests signature byte encoding/decoding error paths
-func TestECDSASignatureBytesErrorPaths(t *testing.T) {
-	// Test decoding with various invalid DER data
-	invalidSignatures := [][]byte{
-		{},                                   // Empty
-		{0x01},                               // Too short
-		{0x31, 0x04, 0x02, 0x01, 0x01},       // Wrong sequence tag
-		{0x30, 0xFF, 0x02, 0x01, 0x01},       // Invalid sequence length
-		{0x30, 0x04, 0x01, 0x01, 0x01},       // Missing INTEGER tag for r
-		{0x30, 0x04, 0x02, 0xFF, 0x01},       // Invalid r length
-		{0x30, 0x06, 0x02, 0x01, 0x01, 0x01}, // Missing INTEGER tag for s
-		{0x30, 0x06, 0x02, 0x01, 0x01, 0x02, 0xFF}, // Invalid s length
-		{0x30, 0x04, 0x02, 0x01},                   // Incomplete data
-	}
-
-	for i, invalidSig := range invalidSignatures {
-		_, _, err := ECDSASignatureFromBytes(invalidSig)
-		if err == nil {
-			t.Errorf("Expected error for invalid signature %d", i)
-		}
-	}
-
-	// Test encoding with edge case values
-	// Test with large numbers that might cause issues
-	largeR := new(big.Int)
-	largeR.SetString("123456789012345678901234567890123456789012345678901234567890", 10)
-	largeS := new(big.Int)
-	largeS.SetString("987654321098765432109876543210987654321098765432109876543210", 10)
-
-	sigBytes, err := ECDSASignatureToBytes(largeR, largeS)
-	if err != nil {
-		t.Errorf("Should handle large signature components: %v", err)
-	}
-
-	decodedR, decodedS, err := ECDSASignatureFromBytes(sigBytes)
-	if err != nil {
-		t.Errorf("Should decode large signature components: %v", err)
-	}
-
-	if decodedR.Cmp(largeR) != 0 || decodedS.Cmp(largeS) != 0 {
-		t.Error("Large signature components should round-trip correctly")
-	}
-
-	// Test with numbers that have high bit set (need 0x00 prefix)
-	highBitR := new(big.Int).SetBytes([]byte{0xFF, 0xFF, 0xFF, 0xFF})
-	highBitS := new(big.Int).SetBytes([]byte{0x80, 0x00, 0x00, 0x01})
-
-	sigBytes2, err := ECDSASignatureToBytes(highBitR, highBitS)
-	if err != nil {
-		t.Errorf("Should handle high-bit signature components: %v", err)
-	}
-
-	decodedR2, decodedS2, err := ECDSASignatureFromBytes(sigBytes2)
-	if err != nil {
-		t.Errorf("Should decode high-bit signature components: %v", err)
-	}
-
-	if decodedR2.Cmp(highBitR) != 0 || decodedS2.Cmp(highBitS) != 0 {
-		t.Error("High-bit signature components should round-trip correctly")
-	}
-}
-
-// TestECDSAEdgeCases tests various edge cases
-func TestECDSAEdgeCases(t *testing.T) {
-	// Test with minimal data
-	keyPair, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate key: %v", err)
-	}
-
-	// Test with empty data
-	r, s, err := ECDSASignSHA256(keyPair.PrivateKey, []byte{})
-	if err != nil {
-		t.Errorf("Should handle empty data: %v", err)
-	}
-
-	if !ECDSAVerifySHA256(keyPair.PublicKey, []byte{}, r, s) {
-		t.Error("Should verify empty data signature")
-	}
-
-	// Test with single byte
-	r2, s2, err := ECDSASignSHA256(keyPair.PrivateKey, []byte{0x42})
-	if err != nil {
-		t.Errorf("Should handle single byte: %v", err)
-	}
-
-	if !ECDSAVerifySHA256(keyPair.PublicKey, []byte{0x42}, r2, s2) {
-		t.Error("Should verify single byte signature")
-	}
-
-	// Test signature round-trip with zero components
-	zeroSig, err := ECDSASignatureToBytes(big.NewInt(0), big.NewInt(1))
-	if err != nil {
-		t.Errorf("Should handle zero R component: %v", err)
-	}
-
-	zeroR, zeroS, err := ECDSASignatureFromBytes(zeroSig)
-	if err != nil {
-		t.Errorf("Should decode zero R component: %v", err)
-	}
-
-	if zeroR.Cmp(big.NewInt(0)) != 0 || zeroS.Cmp(big.NewInt(1)) != 0 {
-		t.Error("Zero component should round-trip correctly")
-	}
-}
-
-// TestECDHEdgeCases tests ECDH edge cases
-func TestECDHEdgeCases(t *testing.T) {
-	keyPair, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate key: %v", err)
-	}
-
-	// Test self key exchange (Alice = Bob)
-	sharedSecret, err := ECDHComputeShared(keyPair.PrivateKey, keyPair.PublicKey)
-	if err != nil {
-		t.Errorf("Self key exchange should work: %v", err)
-	}
-
-	if len(sharedSecret) == 0 {
-		t.Error("Self shared secret should not be empty")
-	}
-
-	// Test KDF with very small key length
-	derivedKey, err := ECDHComputeSharedWithKDF(keyPair.PrivateKey, keyPair.PublicKey, 1, sha256.New)
-	if err != nil {
-		t.Errorf("KDF should work with small key length: %v", err)
-	}
-
-	if len(derivedKey) != 1 {
-		t.Errorf("Expected key length 1, got %d", len(derivedKey))
-	}
-
-	// Test with key length larger than hash output (will trigger counter mode)
-	largeKey, err := ECDHComputeSharedWithKDF(keyPair.PrivateKey, keyPair.PublicKey, 100, sha256.New)
-	if err != nil {
-		t.Errorf("KDF should work with large key length: %v", err)
-	}
-
-	if len(largeKey) != 100 {
-		t.Errorf("Expected key length 100, got %d", len(largeKey))
-	}
-
-	// Test coordinate conversion with curve edge points
-	// Test with generator point
-	gx, gy := elliptic.P256().Params().Gx, elliptic.P256().Params().Gy
-	pubKey, err := ECDHPublicKeyFromCoordinates(elliptic.P256(), gx, gy)
-	if err != nil {
-		t.Errorf("Should handle generator point: %v", err)
-	}
-
-	x, y, err := ECDHPublicKeyToCoordinates(pubKey)
-	if err != nil {
-		t.Errorf("Should extract generator coordinates: %v", err)
-	}
-
-	if x.Cmp(gx) != 0 || y.Cmp(gy) != 0 {
-		t.Error("Generator coordinates should round-trip correctly")
-	}
-}
-
-// TestECCUtilityFunctions tests utility functions thoroughly
-func TestECCUtilityFunctions(t *testing.T) {
-	// Test GetCurveName with all supported curves
-	testCases := []struct {
-		curve elliptic.Curve
-		name  string
-	}{
-		{elliptic.P224(), "P-224"},
-		{elliptic.P256(), "P-256"},
-		{elliptic.P384(), "P-384"},
-		{elliptic.P521(), "P-521"},
-	}
-
-	for _, tc := range testCases {
-		if GetCurveName(tc.curve) != tc.name {
-			t.Errorf("Wrong curve name for %s", tc.name)
-		}
-	}
-
-	// Test IsValidCurve with all cases
-	for _, tc := range testCases {
-		if !IsValidCurve(tc.curve) {
-			t.Errorf("Curve %s should be valid", tc.name)
-		}
-	}
-
-	// Test with nil curve
-	if GetCurveName(nil) != "Unknown" {
-		t.Error("nil curve should return 'Unknown'")
-	}
-
-	if IsValidCurve(nil) {
-		t.Error("nil curve should be invalid")
-	}
-}
-
-// TestECCRealWorldScenarios tests realistic usage scenarios
-func TestECCRealWorldScenarios(t *testing.T) {
-	// Test complete ECDSA workflow
-	// 1. Generate key pair
-	signer, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate signer key: %v", err)
-	}
-
-	// 2. Export keys to PEM
-	privateKeyPEM, err := ECDSAPrivateKeyToPEM(signer.PrivateKey)
-	if err != nil {
-		t.Fatalf("Failed to export private key: %v", err)
-	}
-
-	publicKeyPEM, err := ECDSAPublicKeyToPEM(signer.PublicKey)
-	if err != nil {
-		t.Fatalf("Failed to export public key: %v", err)
-	}
-
-	// 3. Import keys from PEM (simulate key distribution)
-	importedPrivateKey, err := ECDSAPrivateKeyFromPEM(privateKeyPEM)
-	if err != nil {
-		t.Fatalf("Failed to import private key: %v", err)
-	}
-
-	importedPublicKey, err := ECDSAPublicKeyFromPEM(publicKeyPEM)
-	if err != nil {
-		t.Fatalf("Failed to import public key: %v", err)
-	}
-
-	// 4. Sign document
-	document := []byte("Important contract data")
-	r, s, err := ECDSASignSHA256(importedPrivateKey, document)
-	if err != nil {
-		t.Fatalf("Failed to sign document: %v", err)
-	}
-
-	// 5. Export signature
-	sigBytes, err := ECDSASignatureToBytes(r, s)
-	if err != nil {
-		t.Fatalf("Failed to export signature: %v", err)
-	}
-
-	// 6. Import signature (simulate transmission)
-	importedR, importedS, err := ECDSASignatureFromBytes(sigBytes)
-	if err != nil {
-		t.Fatalf("Failed to import signature: %v", err)
-	}
-
-	// 7. Verify signature
-	if !ECDSAVerifySHA256(importedPublicKey, document, importedR, importedS) {
-		t.Error("Failed to verify document signature")
-	}
-
-	// 8. Test with tampered document
-	tamperedDoc := []byte("Tampered contract data")
-	if ECDSAVerifySHA256(importedPublicKey, tamperedDoc, importedR, importedS) {
-		t.Error("Should not verify tampered document")
-	}
-
-	// Test complete ECDH workflow
-	// 1. Generate key pairs for two parties
-	alice, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate Alice's key: %v", err)
-	}
-
-	bob, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate Bob's key: %v", err)
-	}
-
-	// 2. Validate key pairs
-	if err := ValidateECDHKeyPair(alice); err != nil {
-		t.Errorf("Alice's key pair should be valid: %v", err)
-	}
-
-	if err := ValidateECDHKeyPair(bob); err != nil {
-		t.Errorf("Bob's key pair should be valid: %v", err)
-	}
-
-	// 3. Exchange public keys and compute shared secrets
-	aliceShared, err := ECDHKeyExchange(alice.PrivateKey, bob.PublicKey, 32)
-	if err != nil {
-		t.Fatalf("Alice's key exchange failed: %v", err)
-	}
-
-	bobShared, err := ECDHKeyExchange(bob.PrivateKey, alice.PublicKey, 32)
-	if err != nil {
-		t.Fatalf("Bob's key exchange failed: %v", err)
-	}
-
-	// 4. Verify shared secrets match
-	match, err := ECDHSharedSecretTest(alice, bob)
-	if err != nil {
-		t.Errorf("ECDH shared secret test failed: %v", err)
-	}
-	if !match {
-		t.Error("ECDH shared secret test should pass")
-	}
-
-	// 5. Use shared secret as encryption key (simulate)
-	if len(aliceShared) != 32 || len(bobShared) != 32 {
-		t.Error("Shared secrets should be 32 bytes")
-	}
-
-	// Test that different key pairs produce different secrets
-	charlie, _ := GenerateECDHP256Key()
-	charlieShared, _ := ECDHKeyExchange(alice.PrivateKey, charlie.PublicKey, 32)
-
-	if string(aliceShared) == string(charlieShared) {
-		t.Error("Different key pairs should produce different shared secrets")
-	}
-}
-
-// TestECDHMissingCoverage tests specific ECDH error paths to reach 100% coverage
-func TestECDHMissingCoverage(t *testing.T) {
-	// Test ECDHComputeShared with public key not on curve (artificial case)
-	// This will test the curve validation path
-	keyPair1, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate key pair 1: %v", err)
-	}
-
-	keyPair2, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate key pair 2: %v", err)
-	}
-
-	// Test different curves mismatch
-	keyPairP384, err := GenerateECDHP384Key()
-	if err != nil {
-		t.Fatalf("Failed to generate P384 key: %v", err)
-	}
-
-	_, err = ECDHComputeShared(keyPair1.PrivateKey, keyPairP384.PublicKey)
-	if err == nil {
-		t.Error("Expected error for curve mismatch")
-	}
-
-	// Test ValidateECDHKeyPair with mismatched keys
-	// Create key pair with wrong public key
-	wrongKeyPair := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,
-		PublicKey:  keyPair2.PublicKey, // Different public key
-	}
-
-	err = ValidateECDHKeyPair(wrongKeyPair)
-	if err == nil {
-		t.Error("Expected error for mismatched key pair")
-	}
-
-	// Test ECDHSharedSecretTest to ensure it executes without error
-	// We don't assert on the match result since different keys could theoretically match
-	_, err = ECDHSharedSecretTest(keyPair1, keyPair2)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest should not error with valid keys: %v", err)
-	}
-
-	// Test with same key pair (should match)
-	match, err := ECDHSharedSecretTest(keyPair1, keyPair1)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest should not error with same key: %v", err)
-	}
-	if !match {
-		t.Error("Same key pair should have matching secrets")
-	}
-
-	// Test ECDHComputeSharedWithKDF with key length exactly equal to hash length
-	derivedKey, err := ECDHComputeSharedWithKDF(keyPair1.PrivateKey, keyPair2.PublicKey, 32, sha256.New)
-	if err != nil {
-		t.Errorf("KDF should work with key length equal to hash length: %v", err)
-	}
-	if len(derivedKey) != 32 {
-		t.Errorf("Expected key length 32, got %d", len(derivedKey))
-	}
-
-	// Test with key length less than hash length (truncation path)
-	derivedKey2, err := ECDHComputeSharedWithKDF(keyPair1.PrivateKey, keyPair2.PublicKey, 16, sha256.New)
-	if err != nil {
-		t.Errorf("KDF should work with key length less than hash length: %v", err)
-	}
-	if len(derivedKey2) != 16 {
-		t.Errorf("Expected key length 16, got %d", len(derivedKey2))
-	}
-}
-
-// TestPEMEncodingErrorPaths tests the remaining PEM encoding/decoding paths
-func TestPEMEncodingErrorPaths(t *testing.T) {
-	// Generate test keys
-	ecdsaKeyPair, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate ECDSA key pair: %v", err)
-	}
-
-	// Test ECDSAPrivateKeyToPEM success path and internal error handling
-	pemData, err := ECDSAPrivateKeyToPEM(ecdsaKeyPair.PrivateKey)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyToPEM should succeed: %v", err)
-	}
-	if len(pemData) == 0 {
-		t.Error("PEM data should not be empty")
-	}
-
-	// Test ECDSAPublicKeyToPEM success path
-	pubPemData, err := ECDSAPublicKeyToPEM(ecdsaKeyPair.PublicKey)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyToPEM should succeed: %v", err)
-	}
-	if len(pubPemData) == 0 {
-		t.Error("Public PEM data should not be empty")
-	}
-
-	// Test PEM parsing with edge cases that trigger different code paths
-
-	// Test malformed ASN.1 in private key PEM
-	malformedPrivatePEM := `-----BEGIN EC PRIVATE KEY-----
-MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQg
------END EC PRIVATE KEY-----`
-
-	_, err = ECDSAPrivateKeyFromPEM([]byte(malformedPrivatePEM))
-	if err == nil {
-		t.Error("Expected error for malformed private key ASN.1")
-	}
-
-	// Test malformed ASN.1 in public key PEM
-	malformedPublicPEM := `-----BEGIN PUBLIC KEY-----
-MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE
------END PUBLIC KEY-----`
-
-	_, err = ECDSAPublicKeyFromPEM([]byte(malformedPublicPEM))
-	if err == nil {
-		t.Error("Expected error for malformed public key ASN.1")
-	}
-
-	// Test with non-EC algorithms (should fail the type assertion)
-	rsaPublicPEM := `-----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA1hKkprG8j0jLH9s7IqiK
-6VgKtCWNOhd2HWqe1bF/Xy4yU1R1ZRswRWuAHVEJl2Gx4H2Xc7k/lP7I0nN1Q8Xs
-F9Pl5b3r6ZkQXVmV1tTc8GmZ7u4mQ3b1x8cNlmPX7r7a6Q5H8VvT1K2tG3k6w8Kn
-8r7S7Gx1V8R5L5S8s2r5qT7QGk7o5R8L8F2I7d2uV1X7cM4W8K8r7S7K7I1M7V7Q
-7K7J3Q8O7p5f2c4D3b8r7X6K8s7S7Q7K7k7G1V8L8F2I7t4n7w7S4U7o3h6G3s3S3K
-5Q4I3j4K3m7p7N5T7w7R9s7T7K7Q8K7G3s7S7K8n8s7w6h7I4K2Q4K2A6J4K2E5E7X
-2Q3W3E2G3X7N5C7V3K5L8K5Q7K2A5D7P1K5Q3K3A5X3K5N5c2L5T5w7I2I2K8K8K5A
-5D7K7J7w7T4w7r7b3j4k7s5j7k5s7s5v7e7t3e6h2c4n5j7h1j2z4m8h6f7j8i5l7l
-5k7s5j7k5s7s5v7e7t3e6h2c4n5j7h1j2z4m8h6f7j8i5l7l5k7s5j7k5s7s5v7e7t
-3e6h2c4n5j7h1j2z4m8h6f7j8i5l7l5k7s5j7k5s7s5v7e7t3e6h2c4n5j7h1j2z4m
-8h6f7j8i5l7l5k7s5j7k5s7s5v7e7t3e6h2c4n5j7h1j2z4m8h6f7j8i5l7lQIDAQAB
------END PUBLIC KEY-----`
-
-	_, err = ECDSAPublicKeyFromPEM([]byte(rsaPublicPEM))
-	if err == nil {
-		t.Error("Expected error for RSA public key in ECDSA parser")
-	}
-}
-
-// TestECDHSpecificCoveragePaths tests specific code paths to reach 100%
-func TestECDHSpecificCoveragePaths(t *testing.T) {
-	// Generate key pairs for testing
-	keyPair1, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate key pair: %v", err)
-	}
-
-	// Create an invalid public key by modifying coordinates to be off the curve
-	// This should trigger the IsOnCurve check in ECDHComputeShared
-	invalidPublicKey := &ecdsa.PublicKey{
-		Curve: elliptic.P256(),
-		X:     big.NewInt(1), // Invalid point
-		Y:     big.NewInt(1), // Invalid point
-	}
-
-	_, err = ECDHComputeShared(keyPair1.PrivateKey, invalidPublicKey)
-	if err == nil {
-		t.Error("Expected error for public key not on curve")
-	}
-
-	// Create key pair with mismatched curve types to trigger ValidateECDHKeyPair paths
-	keyPairP384, err := GenerateECDHP384Key()
-	if err != nil {
-		t.Fatalf("Failed to generate P384 key pair: %v", err)
-	}
-
-	// Test ValidateECDHKeyPair with curve mismatch
-	mismatchedKeyPair := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,   // P256
-		PublicKey:  keyPairP384.PublicKey, // P384
-	}
-
-	err = ValidateECDHKeyPair(mismatchedKeyPair)
-	if err == nil {
-		t.Error("Expected error for curve mismatch in key pair")
-	}
-
-	// Test ValidateECDHKeyPair with invalid public key coordinates
-	invalidCoordKeyPair := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,
-		PublicKey:  invalidPublicKey,
-	}
-
-	err = ValidateECDHKeyPair(invalidCoordKeyPair)
-	if err == nil {
-		t.Error("Expected error for invalid public key coordinates")
-	}
-}
-
-// TestECDSASpecificCoveragePaths tests ECDSA specific paths for 100% coverage
-func TestECDSASpecificCoveragePaths(t *testing.T) {
-	// Test successful PEM operations to trigger success paths
-	keyPair, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate ECDSA key pair: %v", err)
-	}
-
-	// Test ECDSAPrivateKeyToPEM success path (85.7% -> 100%)
-	privPEM, err := ECDSAPrivateKeyToPEM(keyPair.PrivateKey)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyToPEM should succeed: %v", err)
-	}
-	if len(privPEM) == 0 {
-		t.Error("Private PEM should not be empty")
-	}
-
-	// Test ECDSAPublicKeyToPEM success path (85.7% -> 100%)
-	pubPEM, err := ECDSAPublicKeyToPEM(keyPair.PublicKey)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyToPEM should succeed: %v", err)
-	}
-	if len(pubPEM) == 0 {
-		t.Error("Public PEM should not be empty")
-	}
-
-	// Test ECDSAPrivateKeyFromPEM success path (90.9% -> 100%)
-	decodedPriv, err := ECDSAPrivateKeyFromPEM(privPEM)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPriv == nil {
-		t.Error("Decoded private key should not be nil")
-	}
-
-	// Test ECDSAPublicKeyFromPEM success path (78.6% -> 100%)
-	decodedPub, err := ECDSAPublicKeyFromPEM(pubPEM)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPub == nil {
-		t.Error("Decoded public key should not be nil")
-	}
-
-	// Test with malformed DER signature to trigger ECDSASignatureFromBytes error paths
-
-	// DER with sequence tag but wrong type for first integer
-	malformedDER1 := []byte{0x30, 0x08, 0x04, 0x02, 0x01, 0x01, 0x02, 0x02, 0x01, 0x01}
-	_, _, err = ECDSASignatureFromBytes(malformedDER1)
-	if err == nil {
-		t.Error("Expected error for malformed DER with wrong type")
-	}
-
-	// DER with sequence but truncated after r length
-	malformedDER2 := []byte{0x30, 0x08, 0x02, 0x02, 0x01}
-	_, _, err = ECDSASignatureFromBytes(malformedDER2)
-	if err == nil {
-		t.Error("Expected error for truncated DER after r length")
-	}
-
-	// DER with sequence but wrong type for second integer (s)
-	malformedDER3 := []byte{0x30, 0x08, 0x02, 0x02, 0x01, 0x01, 0x04, 0x02, 0x01, 0x01}
-	_, _, err = ECDSASignatureFromBytes(malformedDER3)
-	if err == nil {
-		t.Error("Expected error for malformed DER with wrong s type")
-	}
-
-	// DER with sequence but truncated after s length
-	malformedDER4 := []byte{0x30, 0x08, 0x02, 0x02, 0x01, 0x01, 0x02, 0x02, 0x01}
-	_, _, err = ECDSASignatureFromBytes(malformedDER4)
-	if err == nil {
-		t.Error("Expected error for truncated DER after s length")
-	}
-
-	// DER with truncated sequence length
-	malformedDER5 := []byte{0x30, 0xFF, 0x02, 0x02, 0x01, 0x01}
-	_, _, err = ECDSASignatureFromBytes(malformedDER5)
-	if err == nil {
-		t.Error("Expected error for truncated sequence")
-	}
-
-	// Test ECDSAPrivateKeyFromPEM with invalid ASN.1 structure
-	invalidASN1PEM := []byte(`-----BEGIN EC PRIVATE KEY-----
-MIGUAgEAMBAGByqGSM49AgEGBSuBBAAKBHQwcgIBAQQgTest
------END EC PRIVATE KEY-----`)
-
-	_, err = ECDSAPrivateKeyFromPEM(invalidASN1PEM)
-	if err == nil {
-		t.Error("Expected error for invalid ASN.1 structure")
-	}
-
-	// Test ECDSAPublicKeyFromPEM with invalid algorithm identifier
-	invalidAlgoIdPEM := []byte(`-----BEGIN PUBLIC KEY-----
-MFMwDQYJKoZIhvcNAQEBBQADQgAwPwIBAAKCAQEAxxx
------END PUBLIC KEY-----`)
-
-	_, err = ECDSAPublicKeyFromPEM(invalidAlgoIdPEM)
-	if err == nil {
-		t.Error("Expected error for invalid algorithm identifier")
-	}
-
-	// Test with completely invalid PEM content
-	invalidPEMContent := []byte(`-----BEGIN EC PRIVATE KEY-----
-invalid base64 content here!!!
------END EC PRIVATE KEY-----`)
-
-	_, err = ECDSAPrivateKeyFromPEM(invalidPEMContent)
-	if err == nil {
-		t.Error("Expected error for invalid PEM content")
-	}
-
-	// Test public key PEM with invalid content
-	invalidPublicPEMContent := []byte(`-----BEGIN PUBLIC KEY-----
-invalid base64 content here!!!
------END PUBLIC KEY-----`)
-
-	_, err = ECDSAPublicKeyFromPEM(invalidPublicPEMContent)
-	if err == nil {
-		t.Error("Expected error for invalid public PEM content")
-	}
-}
-
-// TestECDHAndECDSAMissingCoverage specifically targets the remaining uncovered lines
-func TestECDHAndECDSAMissingCoverage(t *testing.T) {
-	// Generate key pairs for testing
-	keyPair1, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate ECDH key pair 1: %v", err)
-	}
-
-	keyPair2, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate ECDH key pair 2: %v", err)
-	}
-
-	// Test ECDHSharedSecretTest with error path
-	// Create an invalid key pair that will cause ECDHComputeShared to fail
-	invalidPublicKey := &ecdsa.PublicKey{
-		Curve: elliptic.P256(),
-		X:     big.NewInt(0), // Point at infinity (invalid)
-		Y:     big.NewInt(1),
-	}
-
-	invalidKeyPair := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,
-		PublicKey:  invalidPublicKey,
-	}
-
-	// This should trigger the error path in ECDHSharedSecretTest (71.4% -> 100%)
-	_, err = ECDHSharedSecretTest(keyPair1, invalidKeyPair)
-	if err == nil {
-		t.Error("Expected error in ECDHSharedSecretTest with invalid public key")
-	}
-
-	// Test second error path in ECDHSharedSecretTest
-	_, err = ECDHSharedSecretTest(invalidKeyPair, keyPair1)
-	if err == nil {
-		t.Error("Expected error in ECDHSharedSecretTest with invalid key pair as first param")
-	}
-
-	// Test ValidateECDHKeyPair with all error conditions (92.9% -> 100%)
-
-	// Test with nil key pair
-	err = ValidateECDHKeyPair(nil)
-	if err == nil {
-		t.Error("Expected error for nil key pair")
-	}
-
-	// Test with nil private key in key pair
-	keyPairWithNilPrivate := &ECDHKeyPair{
-		PrivateKey: nil,
-		PublicKey:  keyPair1.PublicKey,
-	}
-	err = ValidateECDHKeyPair(keyPairWithNilPrivate)
-	if err == nil {
-		t.Error("Expected error for nil private key")
-	}
-
-	// Test with nil public key in key pair
-	keyPairWithNilPublic := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,
-		PublicKey:  nil,
-	}
-	err = ValidateECDHKeyPair(keyPairWithNilPublic)
-	if err == nil {
-		t.Error("Expected error for nil public key")
-	}
-
-	// Test ValidateECDHKeyPair with the ECDHKeyPair struct method
-	validPair := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,
-		PublicKey:  keyPair1.PublicKey,
-	}
-	err = ValidateECDHKeyPair(validPair)
-	if err != nil {
-		t.Errorf("ValidateECDHKeyPair should succeed with valid pair: %v", err)
-	}
-
-	// Test ECDHComputeSharedWithKDF with all error paths (95.5% -> 100%)
-
-	// Test with 0 key length
-	_, err = ECDHComputeSharedWithKDF(keyPair1.PrivateKey, keyPair2.PublicKey, 0, nil)
-	if err == nil {
-		t.Error("Expected error for zero key length in ECDHComputeSharedWithKDF")
-	}
-
-	// Test with negative key length
-	_, err = ECDHComputeSharedWithKDF(keyPair1.PrivateKey, keyPair2.PublicKey, -1, nil)
-	if err == nil {
-		t.Error("Expected error for negative key length in ECDHComputeSharedWithKDF")
-	}
-
-	// Test with nil KDF function (line 89)
-	_, err = ECDHComputeSharedWithKDF(keyPair1.PrivateKey, keyPair2.PublicKey, 32, nil)
-	if err == nil {
-		t.Error("Expected error for nil KDF function in ECDHComputeSharedWithKDF")
-	}
-
-	// Test with invalid public key (should trigger ECDHComputeShared error)
-	_, err = ECDHComputeSharedWithKDF(keyPair1.PrivateKey, invalidPublicKey, 32, sha256.New)
-	if err == nil {
-		t.Error("Expected error for invalid public key in ECDHComputeSharedWithKDF")
-	}
-
-	// Test counter mode expansion (key length > hash output length)
-	largeKey, err := ECDHComputeSharedWithKDF(keyPair1.PrivateKey, keyPair2.PublicKey, 100, sha256.New)
-	if err != nil {
-		t.Errorf("ECDHComputeSharedWithKDF should work with large key length: %v", err)
-	}
-	if len(largeKey) != 100 {
-		t.Errorf("Expected key length 100, got %d", len(largeKey))
-	}
-
-	// Test ECDHComputeShared with all error paths (90.0% -> 100%)
-
-	// Test with nil private key
-	_, err = ECDHComputeShared(nil, keyPair1.PublicKey)
-	if err == nil {
-		t.Error("Expected error for nil private key in ECDHComputeShared")
-	}
-
-	// Test with nil public key
-	_, err = ECDHComputeShared(keyPair1.PrivateKey, nil)
-	if err == nil {
-		t.Error("Expected error for nil public key in ECDHComputeShared")
-	}
-
-	// Test with invalid public key coordinates (not on curve)
-	_, err = ECDHComputeShared(keyPair1.PrivateKey, invalidPublicKey)
-	if err == nil {
-		t.Error("Expected error for public key not on curve in ECDHComputeShared")
-	}
-
-	// Test curve mismatch
-	keyPairP384, err := GenerateECDHP384Key()
-	if err != nil {
-		t.Fatalf("Failed to generate P384 key pair: %v", err)
-	}
-
-	_, err = ECDHComputeShared(keyPair1.PrivateKey, keyPairP384.PublicKey)
-	if err == nil {
-		t.Error("Expected error for curve mismatch in ECDHComputeShared")
-	}
-
-	// Test successful ECDHComputeShared to hit the success path (line 77-80)
-	sharedSecret, err := ECDHComputeShared(keyPair1.PrivateKey, keyPair2.PublicKey)
-	if err != nil {
-		t.Errorf("ECDHComputeShared should succeed with valid keys: %v", err)
-	}
-	if len(sharedSecret) == 0 {
-		t.Error("Shared secret should not be empty")
-	}
-
-	// Test ECDHSharedSecretTest with nil key pairs (to trigger nil check)
-	_, err = ECDHSharedSecretTest(nil, keyPair1)
-	if err == nil {
-		t.Error("Expected error for nil keyPair1 in ECDHSharedSecretTest")
-	}
-
-	_, err = ECDHSharedSecretTest(keyPair1, nil)
-	if err == nil {
-		t.Error("Expected error for nil keyPair2 in ECDHSharedSecretTest")
-	}
-
-	// Test ECDHSharedSecretTest to trigger the secret length comparison and byte comparison
-	// Create two valid but different key pairs - this should produce different secrets
-	match, err := ECDHSharedSecretTest(keyPair1, keyPair2)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest with valid but different keys should not error: %v", err)
-	}
-	// These should not match since they're different key pairs
-	if match {
-		t.Log("Note: Different key pairs produced matching secrets (unlikely but possible)")
-	}
-
-	// Test ECDHSharedSecretTest to trigger specific byte comparison path
-	// This tests the return false path at line 219
-	keyPair3, _ := GenerateECDHP256Key()
-	keyPair4, _ := GenerateECDHP256Key()
-
-	// These should not match
-	match, err = ECDHSharedSecretTest(keyPair3, keyPair4)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest should not error: %v", err)
-	}
-	// The secrets from different key pairs should not match
-	if match {
-		t.Log("Note: Different key pairs produced matching secrets (very unlikely)")
-	}
-
-	// Test ValidateECDHKeyPair with key pair where public key doesn't match private key
-	// Create a mismatched key pair (same curve, but wrong public key)
-	mismatchedPubKeyPair := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,
-		PublicKey:  keyPair2.PublicKey,
-	}
-	err = ValidateECDHKeyPair(mismatchedPubKeyPair)
-	if err == nil {
-		t.Error("Expected error for public key not matching private key")
-	}
-}
-
-// TestFinalCoverageEdgeCases targets the last remaining uncovered lines
-func TestFinalCoverageEdgeCases(t *testing.T) {
-	// Generate ECDSA key pairs for PEM testing
-	ecdsaKeyPair, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate ECDSA key pair: %v", err)
-	}
-
-	// Test the success path of ECDSAPrivateKeyToPEM to reach 100%
-	privPEM, err := ECDSAPrivateKeyToPEM(ecdsaKeyPair.PrivateKey)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyToPEM should succeed: %v", err)
-	}
-
-	// Test the success path of ECDSAPublicKeyToPEM to reach 100%
-	pubPEM, err := ECDSAPublicKeyToPEM(ecdsaKeyPair.PublicKey)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyToPEM should succeed: %v", err)
-	}
-
-	// Test ECDSA PEM round-trip to trigger all success paths
-	decodedPriv, err := ECDSAPrivateKeyFromPEM(privPEM)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPriv == nil {
-		t.Error("Decoded private key should not be nil")
-	}
-
-	decodedPub, err := ECDSAPublicKeyFromPEM(pubPEM)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPub == nil {
-		t.Error("Decoded public key should not be nil")
-	}
-
-	// Test ECDSASignatureFromBytes with various malformed inputs to trigger all error paths
-
-	// Empty input
-	_, _, err = ECDSASignatureFromBytes([]byte{})
-	if err == nil {
-		t.Error("Expected error for empty signature bytes")
-	}
-
-	// Non-sequence tag
-	malformedDER := []byte{0x04, 0x08, 0x02, 0x02, 0x01, 0x01, 0x02, 0x02, 0x01, 0x01}
-	_, _, err = ECDSASignatureFromBytes(malformedDER)
-	if err == nil {
-		t.Error("Expected error for non-sequence tag")
-	}
-
-	// Too short to read sequence length
-	malformedDER = []byte{0x30}
-	_, _, err = ECDSASignatureFromBytes(malformedDER)
-	if err == nil {
-		t.Error("Expected error for truncated sequence length")
-	}
-
-	// Sequence length longer than remaining bytes
-	malformedDER = []byte{0x30, 0xFF, 0x02, 0x02}
-	_, _, err = ECDSASignatureFromBytes(malformedDER)
-	if err == nil {
-		t.Error("Expected error for sequence length exceeding data")
-	}
-
-	// Test successful signature round-trip to ensure all paths are hit
-	// Create a signature and then decode it to hit success paths
-	r, s, err := ECDSASignSHA256(ecdsaKeyPair.PrivateKey, []byte("test message"))
-	if err != nil {
-		t.Errorf("ECDSASignSHA256 should succeed: %v", err)
-	}
-
-	sigBytes, err := ECDSASignatureToBytes(r, s)
-	if err != nil {
-		t.Errorf("ECDSASignatureToBytes should succeed: %v", err)
-	}
-
-	decodedR, decodedS, err := ECDSASignatureFromBytes(sigBytes)
-	if err != nil {
-		t.Errorf("ECDSASignatureFromBytes should succeed: %v", err)
-	}
-	if decodedR.Cmp(r) != 0 || decodedS.Cmp(s) != 0 {
-		t.Error("Decoded signature should match original")
-	}
-}
-
-// TestComplete100PercentCoverage targets the exact remaining uncovered lines
-func TestComplete100PercentCoverage(t *testing.T) {
-	// Focus on ECDHSharedSecretTest 85.7% -> 100%
-	// We need to ensure ALL paths are executed, especially the return true path
-
-	keyPair1, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate key pair 1: %v", err)
-	}
-
-	// Test with the SAME key pair - this should guarantee return true (line 223)
-	match, err := ECDHSharedSecretTest(keyPair1, keyPair1)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest with identical keys should succeed: %v", err)
-	}
-	if !match {
-		t.Error("Identical key pairs should produce matching secrets")
-	}
-
-	// Create a manually constructed key pair to ensure we hit the success path
-	keyPair2, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate key pair 2: %v", err)
-	}
-
-	// Test ECDHSharedSecretTest success path - computing Alice->Bob and Bob->Alice
-	// This should exercise all the comparison logic (lines 213-223)
-	_, err = ECDHSharedSecretTest(keyPair1, keyPair2)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest should not error with valid different keys: %v", err)
-	}
-
-	// Now test ValidateECDHKeyPair 92.9% -> 100%
-	// We need to ensure the success path (line 161) is hit
-	validKeyPair := &ECDHKeyPair{
-		PrivateKey: keyPair1.PrivateKey,
-		PublicKey:  keyPair1.PublicKey,
-	}
-	err = ValidateECDHKeyPair(validKeyPair)
-	if err != nil {
-		t.Errorf("ValidateECDHKeyPair should succeed with matching key pair: %v", err)
-	}
-
-	// Test ECDHComputeShared 90.0% -> 100%
-	// We need to ensure the success return path (line 80) is covered
-	sharedSecret, err := ECDHComputeShared(keyPair1.PrivateKey, keyPair1.PublicKey)
-	if err != nil {
-		t.Errorf("ECDHComputeShared with valid matching keys should succeed: %v", err)
-	}
-	if len(sharedSecret) == 0 {
-		t.Error("Shared secret should not be empty")
-	}
-
-	// Ensure we test different valid key pairs too
-	sharedSecret2, err := ECDHComputeShared(keyPair1.PrivateKey, keyPair2.PublicKey)
-	if err != nil {
-		t.Errorf("ECDHComputeShared with valid different keys should succeed: %v", err)
-	}
-	if len(sharedSecret2) == 0 {
-		t.Error("Shared secret 2 should not be empty")
-	}
-}
-
-// TestECDSAandRSAPEMComplete targets PEM functions to reach 100%
-func TestECDSAandRSAPEMComplete(t *testing.T) {
-	// Generate keys for comprehensive PEM testing
-	ecdsaKey, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate ECDSA key: %v", err)
-	}
-
-	rsaKey, err := GenerateRSAKeyPair(2048)
-	if err != nil {
-		t.Fatalf("Failed to generate RSA key: %v", err)
-	}
-
-	// Test ALL success paths for ECDSA PEM functions
-
-	// ECDSAPrivateKeyToPEM success path (85.7% -> 100%)
-	ecdsaPrivPEM, err := ECDSAPrivateKeyToPEM(ecdsaKey.PrivateKey)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyToPEM should succeed: %v", err)
-	}
-	if len(ecdsaPrivPEM) == 0 {
-		t.Error("ECDSA private PEM should not be empty")
-	}
-
-	// ECDSAPublicKeyToPEM success path (85.7% -> 100%)
-	ecdsaPubPEM, err := ECDSAPublicKeyToPEM(ecdsaKey.PublicKey)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyToPEM should succeed: %v", err)
-	}
-	if len(ecdsaPubPEM) == 0 {
-		t.Error("ECDSA public PEM should not be empty")
-	}
-
-	// ECDSAPrivateKeyFromPEM success path (90.9% -> 100%)
-	decodedECDSAPriv, err := ECDSAPrivateKeyFromPEM(ecdsaPrivPEM)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyFromPEM should succeed: %v", err)
-	}
-	if decodedECDSAPriv == nil {
-		t.Error("Decoded ECDSA private key should not be nil")
-	}
-
-	// ECDSAPublicKeyFromPEM success path (78.6% -> 100%)
-	decodedECDSAPub, err := ECDSAPublicKeyFromPEM(ecdsaPubPEM)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyFromPEM should succeed: %v", err)
-	}
-	if decodedECDSAPub == nil {
-		t.Error("Decoded ECDSA public key should not be nil")
-	}
-
-	// Test ALL success paths for RSA PEM functions
-
-	// PrivateKeyToPEM success path (85.7% -> 100%)
-	rsaPrivPEM, err := rsaKey.PrivateKeyToPEM()
-	if err != nil {
-		t.Errorf("RSA PrivateKeyToPEM should succeed: %v", err)
-	}
-	if len(rsaPrivPEM) == 0 {
-		t.Error("RSA private PEM should not be empty")
-	}
-
-	// PublicKeyToPEM success path (85.7% -> 100%)
-	rsaPubPEM, err := rsaKey.PublicKeyToPEM()
-	if err != nil {
-		t.Errorf("RSA PublicKeyToPEM should succeed: %v", err)
-	}
-	if len(rsaPubPEM) == 0 {
-		t.Error("RSA public PEM should not be empty")
-	}
-
-	// PrivateKeyFromPEM success path (94.7% -> 100%)
-	decodedRSAPriv, err := PrivateKeyFromPEM(rsaPrivPEM)
-	if err != nil {
-		t.Errorf("RSA PrivateKeyFromPEM should succeed: %v", err)
-	}
-	if decodedRSAPriv == nil {
-		t.Error("Decoded RSA private key should not be nil")
-	}
-
-	// PublicKeyFromPEM success path (94.7% -> 100%)
-	decodedRSAPub, err := PublicKeyFromPEM(rsaPubPEM)
-	if err != nil {
-		t.Errorf("RSA PublicKeyFromPEM should succeed: %v", err)
-	}
-	if decodedRSAPub == nil {
-		t.Error("Decoded RSA public key should not be nil")
-	}
-
-	// Test ECDSASignatureFromBytes final paths (91.7% -> 100%)
-	// Create a valid signature and parse it to hit all success paths
-	r, s, err := ECDSASignSHA256(ecdsaKey.PrivateKey, []byte("test"))
-	if err != nil {
-		t.Fatalf("Failed to create ECDSA signature: %v", err)
-	}
-
-	sigBytes, err := ECDSASignatureToBytes(r, s)
-	if err != nil {
-		t.Fatalf("Failed to encode signature: %v", err)
-	}
-
-	// This should hit the final success path in ECDSASignatureFromBytes
-	parsedR, parsedS, err := ECDSASignatureFromBytes(sigBytes)
-	if err != nil {
-		t.Errorf("ECDSASignatureFromBytes should succeed with valid signature: %v", err)
-	}
-	if parsedR.Cmp(r) != 0 || parsedS.Cmp(s) != 0 {
-		t.Error("Parsed signature should match original")
-	}
-}
-
-// TestFinalPEMCoverageEdgeCases creates specific error conditions to hit uncovered lines
-func TestFinalPEMCoverageEdgeCases(t *testing.T) {
-	// Generate valid keys first
-	ecdsaKey, _ := GenerateECDSAP256Key()
-	rsaKey, _ := GenerateRSAKeyPair(2048)
-
-	// Test ECDSAPrivateKeyToPEM error paths
-	// The 85.7% coverage suggests some error path isn't covered
-	// Try with nil key to trigger line 112 error
-	_, err := ECDSAPrivateKeyToPEM(nil)
-	if err == nil {
-		t.Error("Expected error for nil ECDSA private key")
-	}
-
-	// Test successful path to trigger line 125 return
-	pemData, err := ECDSAPrivateKeyToPEM(ecdsaKey.PrivateKey)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyToPEM should succeed: %v", err)
-	}
-	if len(pemData) == 0 {
-		t.Error("PEM data should not be empty")
-	}
-
-	// Test ECDSAPublicKeyToPEM paths
-	_, err = ECDSAPublicKeyToPEM(nil)
-	if err == nil {
-		t.Error("Expected error for nil ECDSA public key")
-	}
-
-	pemPubData, err := ECDSAPublicKeyToPEM(ecdsaKey.PublicKey)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyToPEM should succeed: %v", err)
-	}
-	if len(pemPubData) == 0 {
-		t.Error("Public PEM data should not be empty")
-	}
-
-	// Test RSA PEM error paths
-	// Test with nil keys
-	_, err = rsaKey.PrivateKeyToPEM()
-	if err != nil {
-		t.Errorf("RSA PrivateKeyToPEM should succeed: %v", err)
-	}
-
-	_, err = rsaKey.PublicKeyToPEM()
-	if err != nil {
-		t.Errorf("RSA PublicKeyToPEM should succeed: %v", err)
-	}
-
-	// Create RSA key pair with nil keys to trigger error paths
-	nilRSAKeyPair := &RSAKeyPair{PrivateKey: nil, PublicKey: nil}
-	_, err = nilRSAKeyPair.PrivateKeyToPEM()
-	if err == nil {
-		t.Error("Expected error for nil RSA private key")
-	}
-
-	_, err = nilRSAKeyPair.PublicKeyToPEM()
-	if err == nil {
-		t.Error("Expected error for nil RSA public key")
-	}
-}
-
-// TestUnusualECDHPaths targets very specific lines in ECDH functions
-func TestUnusualECDHPaths(t *testing.T) {
-	// Try to construct specific scenarios that might hit uncovered lines
-
-	// Create key pairs on different curves to test specific error conditions
-	keyP256, _ := GenerateECDHP256Key()
-	keyP384, _ := GenerateECDHP384Key()
-
-	// Test ECDHComputeShared with edge cases
-	// The 90% coverage suggests 1 line out of ~10 isn't covered
-	// Try self-computation (same key)
-	selfSecret, err := ECDHComputeShared(keyP256.PrivateKey, keyP256.PublicKey)
-	if err != nil {
-		t.Errorf("Self ECDH should work: %v", err)
-	}
-	if len(selfSecret) == 0 {
-		t.Error("Self secret should not be empty")
-	}
-
-	// Try cross-curve to trigger curve mismatch
-	_, err = ECDHComputeShared(keyP256.PrivateKey, keyP384.PublicKey)
-	if err == nil {
-		t.Error("Expected curve mismatch error")
-	}
-
-	// Test ValidateECDHKeyPair with very specific conditions
-	// 92.9% suggests there's a specific line not covered
-
-	// Create a key pair that should be valid
-	validPair := &ECDHKeyPair{
-		PrivateKey: keyP256.PrivateKey,
-		PublicKey:  keyP256.PublicKey,
-	}
-
-	err = ValidateECDHKeyPair(validPair)
-	if err != nil {
-		t.Errorf("Valid key pair should pass validation: %v", err)
-	}
-
-	// Test ECDHSharedSecretTest with specific scenarios
-	// 85.7% suggests several lines aren't covered
-
-	// Test with identical key pairs (should return true)
-	same, err := ECDHSharedSecretTest(keyP256, keyP256)
-	if err != nil {
-		t.Errorf("Same key test should work: %v", err)
-	}
-	if !same {
-		t.Error("Same keys should produce matching secrets")
-	}
-
-	// Create a scenario where secrets have different lengths (unlikely but possible)
-	// This is hard to trigger naturally, so let's just ensure the comparison logic works
-	different1, _ := GenerateECDHP256Key()
-	different2, _ := GenerateECDHP256Key()
-
-	match, err := ECDHSharedSecretTest(different1, different2)
-	if err != nil {
-		t.Errorf("Different key test should work: %v", err)
-	}
-	// The result doesn't matter, we just want to exercise the code paths
-	_ = match
-}
-
-// TestPinpointUncoveredLines targets the exact uncovered lines identified in coverage report
-func TestPinpointUncoveredLines(t *testing.T) {
-	// Target line ecdh.go:72.41,74.3 - curve mismatch in ECDHComputeShared
-	// This line: if privateKey.Curve != publicKey.Curve { return nil, errors.New(...) }
-
-	keyP256, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate P256 key: %v", err)
-	}
-
-	keyP384, err := GenerateECDHP384Key()
-	if err != nil {
-		t.Fatalf("Failed to generate P384 key: %v", err)
-	}
-
-	// This should trigger line 72-74 in ECDHComputeShared
-	_, err = ECDHComputeShared(keyP256.PrivateKey, keyP384.PublicKey)
-	if err == nil {
-		t.Error("Expected curve mismatch error")
-	}
-
-	// Target line ecdh.go:151.57,153.3 - curve mismatch in ValidateECDHKeyPair
-	// This line: if keyPair.PrivateKey.Curve != keyPair.PublicKey.Curve { return errors.New(...) }
-
-	// Create a key pair with mismatched curves
-	mismatchedKeyPair := &ECDHKeyPair{
-		PrivateKey: keyP256.PrivateKey, // P256 curve
-		PublicKey:  keyP384.PublicKey,  // P384 curve
-	}
-
-	// This should trigger line 151-153 in ValidateECDHKeyPair
-	err = ValidateECDHKeyPair(mismatchedKeyPair)
-	if err == nil {
-		t.Error("Expected curve mismatch error in ValidateECDHKeyPair")
-	}
-
-	// Target lines ecdh.go:213.34,215.3 and ecdh.go:218.31,220.4 - ECDHSharedSecretTest comparison paths
-	// Line 213-215: if len(secret1) != len(secret2) { return false, nil }
-	// Line 218-220: if secret1[i] != secret2[i] { return false, nil }
-
-	// To trigger these lines, we need to create a scenario where secrets have different lengths or values
-	// This is tricky because normal ECDH with the same curve should produce same-length secrets
-	// But we can create a scenario using key pairs from different operations
-
-	// Create two different key pairs
-	key1, _ := GenerateECDHP256Key()
-	key2, _ := GenerateECDHP256Key()
-
-	// This should produce different secrets and trigger the byte comparison (lines 217-220)
-	match, err := ECDHSharedSecretTest(key1, key2)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest should work with different keys: %v", err)
-	}
-	// Different keys should typically produce different secrets (return false, nil at line 219)
-	if match {
-		// This is technically possible but very unlikely
-		t.Log("Different keys produced matching secrets (very unlikely but possible)")
-	}
-}
-
-// TestECDSAUncoveredPaths targets ECDSA error paths
-func TestECDSAUncoveredPaths(t *testing.T) {
-	// Let's check more ECDSA uncovered lines by looking at the patterns
-	ecdsaKey, err := GenerateECDSAP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate ECDSA key: %v", err)
-	}
-
-	// Test cases that might trigger x509 marshal errors in ECDSAPrivateKeyToPEM
-	// This is hard to trigger normally, but let's ensure we have success path coverage
-	privPEM, err := ECDSAPrivateKeyToPEM(ecdsaKey.PrivateKey)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyToPEM should succeed: %v", err)
-	}
-	if len(privPEM) == 0 {
-		t.Error("Private PEM should not be empty")
-	}
-
-	// Try to trigger marshal error in ECDSAPublicKeyToPEM
-	pubPEM, err := ECDSAPublicKeyToPEM(ecdsaKey.PublicKey)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyToPEM should succeed: %v", err)
-	}
-	if len(pubPEM) == 0 {
-		t.Error("Public PEM should not be empty")
-	}
-
-	// Test all branch paths in ECDSAPublicKeyFromPEM and ECDSAPrivateKeyFromPEM
-	decodedPriv, err := ECDSAPrivateKeyFromPEM(privPEM)
-	if err != nil {
-		t.Errorf("ECDSAPrivateKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPriv == nil {
-		t.Error("Decoded private key should not be nil")
-	}
-
-	decodedPub, err := ECDSAPublicKeyFromPEM(pubPEM)
-	if err != nil {
-		t.Errorf("ECDSAPublicKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPub == nil {
-		t.Error("Decoded public key should not be nil")
-	}
-}
-
-// TestRSAUncoveredPaths targets remaining RSA error paths
-func TestRSAUncoveredPaths(t *testing.T) {
-	// Generate RSA key to test uncovered paths
-	rsaKey, err := GenerateRSAKeyPair(2048)
-	if err != nil {
-		t.Fatalf("Failed to generate RSA key: %v", err)
-	}
-
-	// Test all success paths in RSA PEM functions to ensure they're covered
-	privPEM, err := rsaKey.PrivateKeyToPEM()
-	if err != nil {
-		t.Errorf("RSA PrivateKeyToPEM should succeed: %v", err)
-	}
-
-	pubPEM, err := rsaKey.PublicKeyToPEM()
-	if err != nil {
-		t.Errorf("RSA PublicKeyToPEM should succeed: %v", err)
-	}
-
-	// Test decoding to ensure all paths are covered
-	decodedPriv, err := PrivateKeyFromPEM(privPEM)
-	if err != nil {
-		t.Errorf("RSA PrivateKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPriv == nil {
-		t.Error("Decoded RSA private key should not be nil")
-	}
-
-	decodedPub, err := PublicKeyFromPEM(pubPEM)
-	if err != nil {
-		t.Errorf("RSA PublicKeyFromPEM should succeed: %v", err)
-	}
-	if decodedPub == nil {
-		t.Error("Decoded RSA public key should not be nil")
-	}
-}
-
-// TestExactUncoveredLines uses very precise approaches to trigger the exact uncovered lines
-func TestExactUncoveredLines(t *testing.T) {
-	// CRITICAL: Let me be very precise about these paths
-
-	// 1. First test ECDHComputeShared curve mismatch (ecdh.go:72.41,74.3)
-	keyP256, err := GenerateECDHP256Key()
-	if err != nil {
-		t.Fatalf("Failed to generate P256 key: %v", err)
-	}
-
-	keyP384, err := GenerateECDHP384Key()
-	if err != nil {
-		t.Fatalf("Failed to generate P384 key: %v", err)
-	}
-
-	// Verify they are different curves
-	if keyP256.PrivateKey.Curve == keyP384.PrivateKey.Curve {
-		t.Fatal("Keys should have different curves")
-	}
-
-	// Create a public key on P256 curve, then manually change its curve to P384
-	// This way the coordinates pass IsOnCurve for P256, but curves don't match
-	fakePublicKey := &ecdsa.PublicKey{
-		Curve: elliptic.P384(),     // Different curve
-		X:     keyP256.PublicKey.X, // Same coordinates (valid on P256)
-		Y:     keyP256.PublicKey.Y, // Same coordinates (valid on P256)
-	}
-
-	// This MUST trigger ecdh.go:72.41,74.3 because coordinates are valid on P256 curve
-	// but the curve fields don't match
-	_, err = ECDHComputeShared(keyP256.PrivateKey, fakePublicKey)
-	if err == nil {
-		t.Fatal("MUST have curve mismatch error - this should NOT pass")
-	}
-	if !strings.Contains(err.Error(), "curve mismatch") {
-		t.Fatalf("Expected curve mismatch error, got: %v", err)
-	}
-	t.Logf("SUCCESS: Triggered curve mismatch in ECDHComputeShared: %v", err)
-
-	// 2. Test ValidateECDHKeyPair curve mismatch (ecdh.go:151.57,153.3)
-	mismatchedKeyPair := &ECDHKeyPair{
-		PrivateKey: keyP256.PrivateKey,
-		PublicKey:  fakePublicKey, // Use the same fake public key
-	}
-
-	// Verify curves are different
-	if mismatchedKeyPair.PrivateKey.Curve == mismatchedKeyPair.PublicKey.Curve {
-		t.Fatal("Mismatched key pair should have different curves")
-	}
-
-	// This MUST trigger ecdh.go:151.57,153.3
-	err = ValidateECDHKeyPair(mismatchedKeyPair)
-	if err == nil {
-		t.Fatal("MUST have curve mismatch error in ValidateECDHKeyPair")
-	}
-	if !strings.Contains(err.Error(), "curve mismatch") {
-		t.Fatalf("Expected curve mismatch error in ValidateECDHKeyPair, got: %v", err)
-	}
-	t.Logf("SUCCESS: Triggered curve mismatch in ValidateECDHKeyPair: %v", err)
-
-	// 3. Test ECDHSharedSecretTest to trigger byte comparison paths
-	// ecdh.go:213.34,215.3 and ecdh.go:218.31,220.4
-
-	// Create multiple key pairs to ensure we get different secrets
-	key1, _ := GenerateECDHP256Key()
-	key2, _ := GenerateECDHP256Key()
-
-	// Try multiple combinations to ensure we hit the comparison paths
-	for i := 0; i < 5; i++ {
-		tempKey, _ := GenerateECDHP256Key()
-		match, err := ECDHSharedSecretTest(key1, tempKey)
-		if err != nil {
-			t.Errorf("ECDHSharedSecretTest should work: %v", err)
-		}
-		// We expect most of these to be false, triggering the comparison logic
-		t.Logf("Attempt %d: keys match = %v", i+1, match)
-	}
-
-	// Force a specific test with known different keys
-	match, err := ECDHSharedSecretTest(key1, key2)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest with different keys failed: %v", err)
-	}
-	t.Logf("Different keys match result: %v", match)
-
-	// Test with same key - this should match and exercise success path
-	match, err = ECDHSharedSecretTest(key1, key1)
-	if err != nil {
-		t.Errorf("ECDHSharedSecretTest with same key failed: %v", err)
-	}
-	if !match {
-		t.Error("Same key should always match")
-	}
-	t.Logf("Same key match result: %v", match)
-}
-
-// TestECDHSharedSecretMismatchPaths specifically targets the uncovered comparison paths
-func TestECDHSharedSecretMismatchPaths(t *testing.T) {
-	// The key insight: ECDHSharedSecretTest computes:
-	// secret1 = ECDHComputeShared(keyPair1.PrivateKey, keyPair2.PublicKey)
-	// secret2 = ECDHComputeShared(keyPair2.PrivateKey, keyPair1.PublicKey)
-	//
-	// To get different results, we need keyPair1 and keyPair2 such that:
-	// keyPair1.PrivateKey * keyPair2.PublicKey != keyPair2.PrivateKey * keyPair1.PublicKey
-	//
-	// This can happen if we create mismatched key pairs
-
-	keyA, _ := GenerateECDHP256Key()
-	keyB, _ := GenerateECDHP256Key()
-	keyC, _ := GenerateECDHP256Key()
-
-	// Create intentionally mismatched pairs:
-	// pair1: privateA with publicB
-	// pair2: privateC with publicA
-	// This should cause: privateA*publicA != privateC*publicB
-	mismatchPair1 := &ECDHKeyPair{
-		PrivateKey: keyA.PrivateKey,
-		PublicKey:  keyB.PublicKey, // Different public key
-	}
-	mismatchPair2 := &ECDHKeyPair{
-		PrivateKey: keyC.PrivateKey,
-		PublicKey:  keyA.PublicKey, // Different public key
-	}
-
-	// Test: this should produce different secrets
-	match, err := ECDHSharedSecretTest(mismatchPair1, mismatchPair2)
-	if err != nil {
-		t.Logf("ECDHSharedSecretTest with mismatched pairs failed: %v", err)
-	} else {
-		if !match {
-			t.Log("SUCCESS: Mismatched key pairs produced different secrets - triggered comparison paths!")
-		} else {
-			t.Log("Mismatched pairs still matched (very unlikely)")
-		}
-	}
-
-	// Try many combinations until we find mismatched secrets
-	for i := 0; i < 50; i++ {
-		k1, _ := GenerateECDHP256Key()
-		k2, _ := GenerateECDHP256Key()
-		k3, _ := GenerateECDHP256Key()
-		k4, _ := GenerateECDHP256Key()
-
-		// Create two completely unrelated key pairs
-		mixed1 := &ECDHKeyPair{PrivateKey: k1.PrivateKey, PublicKey: k2.PublicKey}
-		mixed2 := &ECDHKeyPair{PrivateKey: k3.PrivateKey, PublicKey: k4.PublicKey}
-
-		match, err := ECDHSharedSecretTest(mixed1, mixed2)
-		if err != nil {
-			continue // Skip errors and try next combination
-		}
-		if !match {
-			t.Logf("SUCCESS iteration %d: Found mismatched secrets - triggered comparison paths!", i+1)
-			return // We found what we needed
-		}
-	}
-
-	t.Log("Note: All key combinations produced matching results")
-}
-
-// TestECDHSharedSecretLengthMismatch specifically targets the length mismatch path
-func TestECDHSharedSecretLengthMismatch(t *testing.T) {
-	// To trigger length mismatch (ecdh.go:213.34,215.3), we need secrets of different lengths
-	// Different curves produce different-length secrets:
-	// P256: 32 bytes, P384: 48 bytes, P521: 66 bytes
-
-	// However, ECDHSharedSecretTest will fail if the curves don't match
-	// So we need to create a scenario where ECDHComputeShared returns
-	// different length results through some edge case
-
-	// Let's try a more complex approach: create mixed-curve key pairs
-	// that somehow pass validation but produce different secret lengths
-
-	keyP256, _ := GenerateECDHP256Key()
-	keyP384, _ := GenerateECDHP384Key()
-	keyP521, _ := GenerateECDHP521Key()
-
-	// Try combinations that might work with different curve sizes
-	mixedPairs := []*ECDHKeyPair{
-		{PrivateKey: keyP256.PrivateKey, PublicKey: keyP256.PublicKey}, // 32 bytes
-		{PrivateKey: keyP384.PrivateKey, PublicKey: keyP384.PublicKey}, // 48 bytes
-		{PrivateKey: keyP521.PrivateKey, PublicKey: keyP521.PublicKey}, // 66 bytes
-	}
-
-	// Test combinations that might produce different secret lengths
-	for i, pair1 := range mixedPairs {
-		for j, pair2 := range mixedPairs {
-			if i >= j {
-				continue // Skip same and duplicate combinations
-			}
-
-			match, err := ECDHSharedSecretTest(pair1, pair2)
-			if err != nil {
-				// This is expected for cross-curve operations
-				t.Logf("Expected error for cross-curve test (%d,%d): %v", i, j, err)
-				continue
-			}
-
-			if !match {
-				t.Logf("SUCCESS: Found length or content mismatch for curves (%d,%d)", i, j)
-				return
-			}
-		}
-	}
-
-	// If cross-curve didn't work, try to manipulate the computation manually
-	// by creating custom scenarios that might produce different lengths
-	t.Log("Cross-curve tests completed")
-}
-
-// TestECDSAPEMSpecificErrorPaths covers uncovered error paths in ECDSA PEM functions
-func TestECDSAPEMSpecificErrorPaths(t *testing.T) {
-	// Test ECDSAPrivateKeyFromPEM with invalid PEM block type
-	// Create a valid RSA private key and encode it with RSA PRIVATE KEY block type
-	rsaKey, _ := rsa.GenerateKey(rand.Reader, 1024)
-	rsaKeyDER := x509.MarshalPKCS1PrivateKey(rsaKey)
-	invalidPrivatePEM := pem.EncodeToMemory(&pem.Block{
-		Type:  "RSA PRIVATE KEY",
-		Bytes: rsaKeyDER,
-	})
-
-	_, err := ECDSAPrivateKeyFromPEM(invalidPrivatePEM)
-	if err == nil || !strings.Contains(err.Error(), "invalid PEM block type") {
-		t.Errorf("Expected PEM block type error, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered ECDSAPrivateKeyFromPEM invalid block type error")
-	}
-
-	// Test ECDSAPublicKeyFromPEM with empty PEM data
-	_, err = ECDSAPublicKeyFromPEM([]byte{})
-	if err == nil || !strings.Contains(err.Error(), "PEM data cannot be empty") {
-		t.Errorf("Expected empty PEM data error, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered ECDSAPublicKeyFromPEM empty data error")
-	}
-
-	// Test ECDSAPublicKeyFromPEM with invalid PEM block type
-	// Create a valid RSA public key and encode it with wrong block type
-	rsaPublicKeyDER := x509.MarshalPKCS1PublicKey(&rsaKey.PublicKey)
-	invalidPublicPEM := pem.EncodeToMemory(&pem.Block{
-		Type:  "RSA PUBLIC KEY",
-		Bytes: rsaPublicKeyDER,
-	})
-
-	_, err = ECDSAPublicKeyFromPEM(invalidPublicPEM)
-	if err == nil || !strings.Contains(err.Error(), "invalid PEM block type") {
-		t.Errorf("Expected PEM block type error for public key, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered ECDSAPublicKeyFromPEM invalid block type error")
-	}
-
-	// Test ECDSAPublicKeyFromPEM with valid PEM format but wrong key type (RSA key in PUBLIC KEY format)
-	// Create an RSA key and encode it as PUBLIC KEY format to trigger "not an ECDSA public key" error
-	rsaPublicKeyDER2, _ := x509.MarshalPKIXPublicKey(&rsaKey.PublicKey)
-	rsaPublicKeyPEM := pem.EncodeToMemory(&pem.Block{
-		Type:  "PUBLIC KEY",
-		Bytes: rsaPublicKeyDER2,
-	})
-
-	_, err = ECDSAPublicKeyFromPEM(rsaPublicKeyPEM)
-	if err == nil || !strings.Contains(err.Error(), "not an ECDSA public key") {
-		t.Errorf("Expected 'not an ECDSA public key' error, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered ECDSAPublicKeyFromPEM 'not an ECDSA public key' error")
-	}
-}
-
-// TestRSAPEMErrorPaths covers uncovered error paths in RSA PEM functions
-func TestRSAPEMErrorPaths(t *testing.T) {
-	// Test PrivateKeyFromPEM with ECDSA key in PRIVATE KEY format (not RSA key)
-	ecdsaKey, _ := GenerateECDSAP256Key()
-	ecdsaPrivateKeyDER, _ := x509.MarshalPKCS8PrivateKey(ecdsaKey.PrivateKey)
-	ecdsaPrivateKeyPEM := pem.EncodeToMemory(&pem.Block{
-		Type:  "PRIVATE KEY",
-		Bytes: ecdsaPrivateKeyDER,
-	})
-
-	_, err := PrivateKeyFromPEM(ecdsaPrivateKeyPEM)
-	if err == nil || !strings.Contains(err.Error(), "key is not an RSA private key") {
-		t.Errorf("Expected 'key is not an RSA private key' error, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered PrivateKeyFromPEM 'key is not an RSA private key' error")
-	}
-
-	// Test PublicKeyFromPEM with ECDSA key in PUBLIC KEY format (not RSA key)
-	ecdsaPublicKeyDER, _ := x509.MarshalPKIXPublicKey(ecdsaKey.PublicKey)
-	ecdsaPublicKeyPEM := pem.EncodeToMemory(&pem.Block{
-		Type:  "PUBLIC KEY",
-		Bytes: ecdsaPublicKeyDER,
-	})
-
-	_, err = PublicKeyFromPEM(ecdsaPublicKeyPEM)
-	if err == nil || !strings.Contains(err.Error(), "key is not an RSA public key") {
-		t.Errorf("Expected 'key is not an RSA public key' error, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered PublicKeyFromPEM 'key is not an RSA public key' error")
-	}
-}
-
-// TestECDSASignatureParsingErrors covers uncovered error paths in ECDSA signature parsing
-func TestECDSASignatureParsingErrors(t *testing.T) {
-	// Test ECDSASignatureFromBytes with incorrect r length (ecdsa.go:256.24,258.3)
-	// Create malformed DER signature: sequence says correct length but r length field is wrong
-	malformedRLen := []byte{
-		0x30, 0x04, // SEQUENCE of 4 bytes (matches the data that follows)
-		0x02, 0x03, // INTEGER tag claiming length 3
-		0x01, 0x02, // Only 2 bytes available (but r length claims 3)
-	}
-
-	_, _, err := ECDSASignatureFromBytes(malformedRLen)
-	if err == nil || !strings.Contains(err.Error(), "incorrect r length") {
-		t.Errorf("Expected 'incorrect r length' error, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered ECDSASignatureFromBytes 'incorrect r length' error")
-	}
-
-	// Test ECDSASignatureFromBytes with incorrect s length (ecdsa.go:270.24,272.3)
-	// Create malformed DER signature: r is correct but s length field is wrong
-	malformedSLen := []byte{
-		0x30, 0x07, // SEQUENCE of 7 bytes (matches actual data length)
-		0x02, 0x02, 0x01, 0x02, // r: INTEGER with 2 bytes [0x01, 0x02] (correct)
-		0x02, 0x02, // s: INTEGER tag claiming length 2
-		0x03, // Only 1 byte available (but s length claims 2)
-	}
-
-	_, _, err = ECDSASignatureFromBytes(malformedSLen)
-	if err == nil || !strings.Contains(err.Error(), "incorrect s length") {
-		t.Errorf("Expected 'incorrect s length' error, got: %v", err)
-	} else {
-		t.Log("SUCCESS: Triggered ECDSASignatureFromBytes 'incorrect s length' error")
-	}
-}
-
-// TestX509MarshalErrorPaths attempts to trigger x509 marshal errors
-func TestX509MarshalErrorPaths(t *testing.T) {
-	// Attempt 1: Create an ECDSA key with corrupted/nil internal fields that might cause marshal to fail
-	key, _ := GenerateECDSAP256Key()
-
-	// Try to create a corrupted key that passes basic nil checks but fails marshaling
-	corruptedPrivateKey := &ecdsa.PrivateKey{
-		PublicKey: ecdsa.PublicKey{
-			Curve: key.PrivateKey.PublicKey.Curve,
-			X:     key.PrivateKey.PublicKey.X,
-			Y:     key.PrivateKey.PublicKey.Y,
-		},
-		D: big.NewInt(0), // Zero private key might cause marshal issues
-	}
-
-	// Test ECDSAPrivateKeyToPEM with corrupted key (ecdsa.go:116.16,118.3)
-	_, err := ECDSAPrivateKeyToPEM(corruptedPrivateKey)
-	if err != nil && strings.Contains(err.Error(), "failed to marshal private key") {
-		t.Log("SUCCESS: Triggered ECDSAPrivateKeyToPEM marshal error")
-	} else {
-		t.Logf("Corrupted private key marshal result: %v", err)
-	}
-
-	// Try with invalid coordinate public key for marshal error
-	// Use coordinates that are off the curve
-	corruptedPublicKey := &ecdsa.PublicKey{
-		Curve: key.PublicKey.Curve,
-		X:     big.NewInt(1), // Invalid coordinate (not on curve)
-		Y:     big.NewInt(1), // Invalid coordinate (not on curve)
-	}
-
-	// Test ECDSAPublicKeyToPEM with corrupted key (ecdsa.go:158.16,160.3)
-	_, err = ECDSAPublicKeyToPEM(corruptedPublicKey)
-	if err != nil && strings.Contains(err.Error(), "failed to marshal public key") {
-		t.Log("SUCCESS: Triggered ECDSAPublicKeyToPEM marshal error")
-	} else {
-		t.Logf("Corrupted public key marshal result: %v", err)
-	}
-
-	// Test RSA marshal errors using similar corrupted key approach
-	rsaKey, _ := GenerateRSAKeyPair(2048)
-
-	// Corrupted RSA private key with nil fields that might cause marshal to fail
-	corruptedRSAPrivate := &rsa.PrivateKey{
-		PublicKey: rsa.PublicKey{
-			N: rsaKey.PrivateKey.PublicKey.N,
-			E: rsaKey.PrivateKey.PublicKey.E,
-		},
-		D:      big.NewInt(0),                            // Zero private key
-		Primes: []*big.Int{big.NewInt(1), big.NewInt(1)}, // Invalid primes
-	}
-
-	corruptedRSAKeyPair := &RSAKeyPair{
-		PrivateKey: corruptedRSAPrivate,
-		PublicKey:  &corruptedRSAPrivate.PublicKey,
-	}
-
-	// Test RSA PrivateKeyToPEM with corrupted key (rsa.go:59.16,61.3)
-	_, err = corruptedRSAKeyPair.PrivateKeyToPEM()
-	if err != nil && strings.Contains(err.Error(), "failed to marshal private key") {
-		t.Log("SUCCESS: Triggered RSA PrivateKeyToPEM marshal error")
-	} else {
-		t.Logf("Corrupted RSA private key marshal result: %v", err)
-	}
-
-	// Corrupted RSA public key
-	corruptedRSAPublic := &rsa.PublicKey{
-		N: big.NewInt(1), // Invalid small modulus that might cause marshal issues
-		E: rsaKey.PublicKey.E,
-	}
-
-	corruptedRSAPublicKeyPair := &RSAKeyPair{
-		PrivateKey: rsaKey.PrivateKey,
-		PublicKey:  corruptedRSAPublic,
-	}
-
-	// Test RSA PublicKeyToPEM with corrupted key (rsa.go:78.16,80.3)
-	_, err = corruptedRSAPublicKeyPair.PublicKeyToPEM()
-	if err != nil && strings.Contains(err.Error(), "failed to marshal public key") {
-		t.Log("SUCCESS: Triggered RSA PublicKeyToPEM marshal error")
-	} else {
-		t.Logf("Corrupted RSA public key marshal result: %v", err)
-	}
-}
-
-/*
-// TestRemainingMarshalErrors attempts to trigger remaining marshal errors
-// DISABLED due to panic with extreme values
-func TestRemainingMarshalErrorsDisabled(t *testing.T) {
-	// This test causes panics with extreme BigInt values
-	// The remaining marshal errors may be unreachable in practice
-}
-*/
-
-// TestUltimateECDHLengthMismatch attempts to create ECDH length mismatch using extreme techniques
-func TestUltimateECDHLengthMismatch(t *testing.T) {
-	// The ECDH length mismatch (ecdh.go:213.34,215.3) is extremely difficult because:
-	// 1. Same curve operations should produce same length results
-	// 2. The X coordinate byte representation should be consistent
-
-	// Attempt: Try to create a scenario where ECDH computation might produce
-	// different length results by manipulating the BigInt representation
-
-	// Create custom ECDHSharedSecretTest-like function that we can manipulate
-	testCustomECDH := func(secret1, secret2 []byte) bool {
-		// This replicates the exact logic from ECDHSharedSecretTest
-		if len(secret1) != len(secret2) {
-			t.Log("SUCCESS: Achieved length mismatch in ECDH secrets!")
-			return false // This would trigger ecdh.go:213.34,215.3
-		}
-
-		for i := 0; i < len(secret1); i++ {
-			if secret1[i] != secret2[i] {
-				return false
-			}
-		}
-		return true
-	}
-
-	// Try to create scenarios with different BigInt byte representations
-	key1, _ := GenerateECDHP256Key()
-	key2, _ := GenerateECDHP256Key()
-
-	// Get normal shared secrets
-	secret1, _ := ECDHComputeShared(key1.PrivateKey, key2.PublicKey)
-	secret2, _ := ECDHComputeShared(key2.PrivateKey, key1.PublicKey)
-
-	t.Logf("Normal ECDH secret lengths: secret1=%d, secret2=%d", len(secret1), len(secret2))
-
-	// Try multiple different key combinations to see if we can get length differences
-	for i := 0; i < 100; i++ {
-		tempKey1, _ := GenerateECDHP256Key()
-		tempKey2, _ := GenerateECDHP256Key()
-
-		tempSecret1, _ := ECDHComputeShared(tempKey1.PrivateKey, tempKey2.PublicKey)
-		tempSecret2, _ := ECDHComputeShared(tempKey2.PrivateKey, tempKey1.PublicKey)
-
-		if len(tempSecret1) != len(tempSecret2) {
-			t.Logf("Found length mismatch at iteration %d: %d vs %d", i, len(tempSecret1), len(tempSecret2))
-			testCustomECDH(tempSecret1, tempSecret2)
-			return
-		}
-
-		// Also try with artificially manipulated secrets
-		if len(tempSecret1) > 1 {
-			// Try truncating one secret to create length mismatch
-			truncatedSecret := tempSecret1[:len(tempSecret1)-1]
-			if len(truncatedSecret) != len(tempSecret2) {
-				testCustomECDH(truncatedSecret, tempSecret2)
-				// This wouldn't be realistic, but let's see if we can trigger the path
-			}
-		}
-	}
-
-	// Alternative approach: Try to modify the ECDH calculation to force length differences
-	// by creating keys with edge-case coordinates that might produce shorter byte arrays
-
-	t.Log("Attempting to create artificial length mismatch scenarios...")
-
-	// Create keys where the X coordinate might have leading zeros
-	// This could potentially result in shorter byte arrays
-	attemptCount := 0
-	for attemptCount < 50 {
-		k1, _ := GenerateECDHP256Key()
-		k2, _ := GenerateECDHP256Key()
-
-		s1, _ := ECDHComputeShared(k1.PrivateKey, k2.PublicKey)
-		s2, _ := ECDHComputeShared(k2.PrivateKey, k1.PublicKey)
-
-		// Check if X coordinates have different leading zero patterns
-		if len(s1) != len(s2) {
-			t.Logf("SUCCESS: Found natural length mismatch - s1=%d, s2=%d", len(s1), len(s2))
-			testCustomECDH(s1, s2)
-
-			// Now test with the real ECDHSharedSecretTest function
-			mixedPair1 := &ECDHKeyPair{PrivateKey: k1.PrivateKey, PublicKey: k2.PublicKey}
-			mixedPair2 := &ECDHKeyPair{PrivateKey: k2.PrivateKey, PublicKey: k1.PublicKey}
-
-			match, err := ECDHSharedSecretTest(mixedPair1, mixedPair2)
-			if err != nil {
-				t.Logf("ECDH test error: %v", err)
-			} else {
-				t.Logf("ECDH test result with length mismatch: %v", match)
-			}
-			return
-		}
-		attemptCount++
-	}
-
-	t.Log("Could not create natural length mismatch scenario")
-}
-
-// TestDirectECDHLengthMismatch attempts to directly trigger the length mismatch path
-func TestDirectECDHLengthMismatch(t *testing.T) {
-	// Since natural ECDH operations always produce same-length results,
-	// let's try some extreme edge cases
-
-	for attempt := 0; attempt < 100; attempt++ {
-		// Generate completely random key pairs
-		pair1, _ := GenerateECDHP256Key()
-		pair2, _ := GenerateECDHP256Key()
-
-		// Create mixed pairs
-		mixedPair1 := &ECDHKeyPair{
-			PrivateKey: pair1.PrivateKey,
-			PublicKey:  pair2.PublicKey,
-		}
-		mixedPair2 := &ECDHKeyPair{
-			PrivateKey: pair2.PrivateKey,
-			PublicKey:  pair1.PublicKey,
-		}
-
-		// Get the secrets directly to check their lengths
-		secret1, err1 := ECDHComputeShared(mixedPair1.PrivateKey, mixedPair2.PublicKey)
-		secret2, err2 := ECDHComputeShared(mixedPair2.PrivateKey, mixedPair1.PublicKey)
-
-		if err1 != nil || err2 != nil {
-			continue
-		}
-
-		// Check if we have different lengths (very unlikely but possible with BigInt.Bytes())
-		if len(secret1) != len(secret2) {
-			t.Logf("BREAKTHROUGH: Found natural length mismatch at attempt %d", attempt)
-			t.Logf("Secret 1 length: %d, Secret 2 length: %d", len(secret1), len(secret2))
-
-			// Now test with ECDHSharedSecretTest to trigger the actual path
-			result, err := ECDHSharedSecretTest(mixedPair1, mixedPair2)
-			if err != nil {
-				t.Logf("ECDHSharedSecretTest error: %v", err)
-			} else {
-				t.Logf("ECDHSharedSecretTest result: %v", result)
-			}
-
-			// This should have triggered ecdh.go:213.34,215.3
-			t.Log("SUCCESS: Should have triggered ECDH length mismatch path!")
-			return
-		}
-	}
-
-	t.Log("No natural length mismatch found after attempts")
-}
-
-// TestAdvancedMarshalErrorTriggers attempts sophisticated approaches to trigger x509 marshal errors
-func TestAdvancedMarshalErrorTriggers(t *testing.T) {
-	// Approach 1: Try to create ECDSA keys with invalid curve parameters
-	// that pass basic validation but fail during x509 marshaling
-
-	// Create a private key with nil curve (this should be caught earlier, but let's try)
-	invalidPrivateKey := &ecdsa.PrivateKey{
-		PublicKey: ecdsa.PublicKey{
-			Curve: nil, // This might cause x509 marshal to fail
-			X:     big.NewInt(1),
-			Y:     big.NewInt(1),
-		},
-		D: big.NewInt(1),
-	}
-
-	// This might trigger ecdsa.go:116.16,118.3 (marshal error)
-	_, err := ECDSAPrivateKeyToPEM(invalidPrivateKey)
-	if err != nil {
-		t.Logf("Successfully triggered ECDSAPrivateKeyToPEM marshal error with nil curve: %v", err)
-	}
-
-	// Approach 2: Try with zero coordinates that might pass IsOnCurve but fail marshaling
-	curve := elliptic.P256()
-	zeroPrivateKey := &ecdsa.PrivateKey{
-		PublicKey: ecdsa.PublicKey{
-			Curve: curve,
-			X:     big.NewInt(0), // Zero coordinates
-			Y:     big.NewInt(0),
-		},
-		D: big.NewInt(0), // Zero private key
-	}
-
-	_, err = ECDSAPrivateKeyToPEM(zeroPrivateKey)
-	if err != nil {
-		t.Logf("Successfully triggered ECDSAPrivateKeyToPEM marshal error with zero values: %v", err)
-	}
-
-	// Approach 3: Try with RSA public key with invalid parameters
-	// Create RSA key with zero modulus
-	invalidRSAPublicKey := &rsa.PublicKey{
-		N: big.NewInt(0), // Zero modulus should cause marshal issues
-		E: 0,             // Zero exponent
-	}
-
-	// This might trigger rsa.go:78.16,80.3 (marshal error)
-	kp := &RSAKeyPair{PublicKey: invalidRSAPublicKey}
-	_, err = kp.PublicKeyToPEM()
-	if err != nil {
-		t.Logf("Successfully triggered RSA PublicKeyToPEM marshal error with zero values: %v", err)
-	}
-
-	// Approach 4: Try with negative values
-	negativeRSAPublicKey := &rsa.PublicKey{
-		N: big.NewInt(-1), // Negative modulus
-		E: -1,             // Negative exponent
-	}
-
-	kp2 := &RSAKeyPair{PublicKey: negativeRSAPublicKey}
-	_, err = kp2.PublicKeyToPEM()
-	if err != nil {
-		t.Logf("Successfully triggered RSA PublicKeyToPEM marshal error with negative values: %v", err)
-	}
-
-	// Approach 5: Try with extremely large RSA modulus that might cause marshal issues
-	// Create a BigInt that's too large or has invalid properties
-	hugeN := new(big.Int)
-	// Set to a value that might cause x509 marshaling issues
-	hugeN.SetString("115792089210356248762697446949407573530086143415290314195533631308867097853951", 10)
-
-	extremeRSAPublicKey := &rsa.PublicKey{
-		N: hugeN,
-		E: 999999999, // Extremely large exponent
-	}
-
-	kp3 := &RSAKeyPair{PublicKey: extremeRSAPublicKey}
-	_, err = kp3.PublicKeyToPEM()
-	if err != nil {
-		t.Logf("Successfully triggered RSA PublicKeyToPEM marshal error with extreme values: %v", err)
-	}
-
-	// Approach 6: Try to create key pairs with different coordinate byte lengths for ECDH
-	// Generate multiple key pairs and try to find ones that produce different secret lengths
-	for attempt := 1; attempt <= 50; attempt++ {
-		keyPair1, err := GenerateECDHP256Key()
-		if err != nil {
-			continue
-		}
-
-		keyPair2, err := GenerateECDHP384Key() // Different curve - this will cause error but we want secret length diff
-		if err != nil {
-			continue
-		}
-
-		// Try forcing coordinate manipulation to get different byte lengths
-		// Create a modified public key with minimal X coordinate
-		minPublicKey := &ecdsa.PublicKey{
-			Curve: keyPair1.PublicKey.Curve,
-			X:     big.NewInt(1), // Minimal X coordinate
-			Y:     keyPair1.PublicKey.Y,
-		}
-
-		// Check if this is on the curve
-		if !keyPair1.PrivateKey.Curve.IsOnCurve(minPublicKey.X, minPublicKey.Y) {
-			// Try to find a valid minimal point
-			continue
-		}
-
-		// Create artificial keyPairs with manipulated secrets
-		artificialPair1 := &ECDHKeyPair{
-			PrivateKey: keyPair1.PrivateKey,
-			PublicKey:  keyPair1.PublicKey,
-		}
-
-		artificialPair2 := &ECDHKeyPair{
-			PrivateKey: keyPair2.PrivateKey,
-			PublicKey:  minPublicKey, // Use minimal public key
-		}
-
-		// Test the shared secret test - this should trigger some comparison path
-		result, err := ECDHSharedSecretTest(artificialPair1, artificialPair2)
-		if err != nil {
-			// Good, we got an error which means we may have triggered some path
-			t.Logf("Attempt %d: ECDHSharedSecretTest error (expected): %v", attempt, err)
-		} else {
-			t.Logf("Attempt %d: ECDHSharedSecretTest result: %v", attempt, result)
-		}
-
-		if attempt == 50 {
-			t.Log("Completed 50 attempts for ECDH path coverage")
-		}
-	}
-}
-
-// TestFinalRSAMarshalError attempts the most extreme approaches to trigger RSA marshal error
-func TestFinalRSAMarshalError(t *testing.T) {
-	// Approach 1: Try to create a malformed RSA key using reflection
-	malformedKey := &rsa.PublicKey{}
-
-	// Use reflection to set invalid internal fields
-	v := reflect.ValueOf(malformedKey).Elem()
-
-	// Set N field to a malformed BigInt
-	nField := v.FieldByName("N")
-	if nField.IsValid() && nField.CanSet() {
-		malformedBig := &big.Int{}
-		// Try to create a BigInt with invalid internal state using unsafe operations
-		malformedBig.SetString("1", 10)
-
-		// Now try to corrupt it
-		bigIntValue := reflect.ValueOf(malformedBig).Elem()
-		if wordsField := bigIntValue.FieldByName("abs"); wordsField.IsValid() {
-			// Try to create an invalid internal state
-			t.Logf("BigInt internal structure access attempt")
-		}
-
-		nField.Set(reflect.ValueOf(malformedBig))
-	}
-
-	// Set E to an extreme value
-	eField := v.FieldByName("E")
-	if eField.IsValid() && eField.CanSet() {
-		eField.SetInt(0x7FFFFFFF) // Max int value
-	}
-
-	kp := &RSAKeyPair{PublicKey: malformedKey}
-	_, err := kp.PublicKeyToPEM()
-	if err != nil {
-		t.Logf("Successfully triggered RSA PublicKeyToPEM marshal error with reflection: %v", err)
-	}
-
-	// Approach 2: Try with a BigInt that has extreme bit length
-	extremeBig := big.NewInt(1)
-	extremeBig.Lsh(extremeBig, 65536) // Shift left by 65536 bits - extremely large number
-
-	extremeKey := &rsa.PublicKey{
-		N: extremeBig,
-		E: 3,
-	}
-
-	kp2 := &RSAKeyPair{PublicKey: extremeKey}
-	_, err = kp2.PublicKeyToPEM()
-	if err != nil {
-		t.Logf("Successfully triggered RSA PublicKeyToPEM marshal error with extreme bit length: %v", err)
-	}
-
-	// Approach 3: Try with nil BigInt pointer (unsafe)
-	nilBigKey := &rsa.PublicKey{
-		N: nil, // This should definitely cause marshal issues
-		E: 65537,
-	}
-
-	kp3 := &RSAKeyPair{PublicKey: nilBigKey}
-	_, err = kp3.PublicKeyToPEM()
-	if err != nil {
-		t.Logf("Successfully triggered RSA PublicKeyToPEM marshal error with nil BigInt: %v", err)
-	}
-
-	// Approach 4: Try creating a key with malformed internal BigInt structure using unsafe
-	corruptedBig := big.NewInt(1)
-
-	// Use unsafe to corrupt the BigInt's internal structure
-	bigIntPtr := (*struct {
-		neg bool
-		abs []big.Word
-	})(unsafe.Pointer(corruptedBig))
-
-	// Try to create an invalid internal state
-	if bigIntPtr != nil {
-		// Set abs to nil slice which might cause marshal issues
-		bigIntPtr.abs = nil
-
-		corruptedKey := &rsa.PublicKey{
-			N: corruptedBig,
-			E: 65537,
-		}
-
-		kp4 := &RSAKeyPair{PublicKey: corruptedKey}
-		_, err = kp4.PublicKeyToPEM()
-		if err != nil {
-			t.Logf("Successfully triggered RSA PublicKeyToPEM marshal error with corrupted BigInt: %v", err)
-		}
-	}
-}
