--- cryptox/missing_coverage_test.go
@@ -1,150 +1,0 @@
-package cryptox
-
-import (
-	"testing"
-)
-
-// TestECDHSharedSecretTestMissingPaths tests the specific missing paths in ECDHSharedSecretTest
-func TestECDHSharedSecretTestMissingPaths(t *testing.T) {
-	t.Run("force_secret_length_mismatch", func(t *testing.T) {
-		// Create two different key pairs with different curves to potentially get different secret lengths
-		p256Key, err := GenerateECDHP256Key()
-		if err != nil {
-			t.Fatal(err)
-		}
-		
-		p384Key, err := GenerateECDHP384Key()
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// Try to create a scenario where the shared secrets have different lengths
-		// This is tricky because ECDH normally produces secrets of the same length for the same curve
-		
-		// Create a modified key pair that might produce different behavior
-		// We'll create keys from the same curve but try to manipulate the calculation
-		
-		// Actually, let's try a simpler approach - create custom ECDHKeyPair structs
-		// with keys that will fail computation in different ways to get different secret lengths
-		
-		// Generate valid keys first
-		key1, _ := GenerateECDHP256Key()
-		key2, _ := GenerateECDHP256Key()
-		
-		// Test the normal case first
-		match, err := ECDHSharedSecretTest(key1, key2)
-		if err != nil {
-			t.Logf("Normal case error (expected for different keys): %v", err)
-		} else {
-			t.Logf("Normal case match: %v", match)
-		}
-		
-		// Now let's try to create a case where ECDHComputeShared might return different lengths
-		// by using keys with different curves (this should cause an error, but let's test it)
-		
-		// Create mixed curve scenario
-		mixedPair1 := &ECDHKeyPair{
-			PrivateKey: p256Key.PrivateKey,
-			PublicKey:  p384Key.PublicKey, // Different curve!
-		}
-		
-		mixedPair2 := &ECDHKeyPair{
-			PrivateKey: p384Key.PrivateKey,
-			PublicKey:  p256Key.PublicKey, // Different curve!
-		}
-		
-		// This should trigger an error path in ECDHComputeShared
-		match, err = ECDHSharedSecretTest(mixedPair1, mixedPair2)
-		if err != nil {
-			t.Logf("Mixed curve error (expected): %v", err)
-		} else {
-			t.Logf("Mixed curve match: %v", match)
-		}
-	})
-
-	t.Run("force_byte_mismatch", func(t *testing.T) {
-		// Try to create a scenario where the bytes don't match
-		// This is actually the normal case when using different key pairs
-		
-		key1, _ := GenerateECDHP256Key()
-		key2, _ := GenerateECDHP256Key()
-		key3, _ := GenerateECDHP256Key()
-		
-		// Create mismatched pairs - Alice's private with Bob's public, vs Charlie's private with Alice's public
-		mismatchPair1 := &ECDHKeyPair{
-			PrivateKey: key1.PrivateKey,
-			PublicKey:  key2.PublicKey,
-		}
-		
-		mismatchPair2 := &ECDHKeyPair{
-			PrivateKey: key3.PrivateKey,
-			PublicKey:  key1.PublicKey,
-		}
-		
-		// These should not produce matching secrets
-		match, err := ECDHSharedSecretTest(mismatchPair1, mismatchPair2)
-		if err != nil {
-			t.Logf("Mismatched pairs error: %v", err)
-		} else {
-			if match {
-				t.Error("Expected mismatch, but got match")
-			} else {
-				t.Log("Successfully triggered byte mismatch path")
-			}
-		}
-	})
-
-	t.Run("manual_secret_comparison", func(t *testing.T) {
-		// Let's manually test the secret comparison logic by creating
-		// a scenario that should trigger the byte-by-byte comparison
-		
-		key1, _ := GenerateECDHP256Key()
-		key2, _ := GenerateECDHP256Key()
-		
-		// Create proper key pairs for ECDH
-		pair1 := &ECDHKeyPair{
-			PrivateKey: key1.PrivateKey,
-			PublicKey:  key2.PublicKey,
-		}
-		
-		pair2 := &ECDHKeyPair{
-			PrivateKey: key2.PrivateKey,
-			PublicKey:  key1.PublicKey,
-		}
-		
-		// These should produce matching secrets (proper ECDH)
-		match, err := ECDHSharedSecretTest(pair1, pair2)
-		if err != nil {
-			t.Logf("ECDH error (this might happen with mixed key pairs): %v", err)
-		} else {
-			if match {
-				t.Log("Successfully tested matching ECDH comparison")
-			} else {
-				t.Log("Successfully tested non-matching ECDH comparison")
-			}
-		}
-	})
-}
-
-// TestECDHSharedSecretTestNilCheck tests nil parameter handling
-func TestECDHSharedSecretTestNilCheck(t *testing.T) {
-	key1, _ := GenerateECDHP256Key()
-	
-	// Test nil first parameter
-	_, err := ECDHSharedSecretTest(nil, key1)
-	if err == nil {
-		t.Error("Expected error for nil first parameter")
-	}
-	
-	// Test nil second parameter
-	_, err = ECDHSharedSecretTest(key1, nil)
-	if err == nil {
-		t.Error("Expected error for nil second parameter")
-	}
-	
-	// Test both nil
-	_, err = ECDHSharedSecretTest(nil, nil)
-	if err == nil {
-		t.Error("Expected error for both nil parameters")
-	}
-}
